feat: improve DX and safety across APIs - Add assert guards in IContext for API accessors - Add MIDI range validation in MidiAPI with warnings - Change SubscriptionID from uint8_t to uint16_t (255 -> 65535 limit) - Add critical warning when default context fails to initialize - Unify ButtonAPI with EncoderAPI by adding hw_ reference

+++ src/oc/core/event/IEventBus.hpp
@@ -7,7 +7,7 @@
 
 namespace oc::core::event {
 
-using SubscriptionID = uint8_t;
+using SubscriptionID = uint16_t;
 using EventCallback = std::function<void(const Event&)>;
 
 /**
